requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.5.1
black
flake8
isort
sqlalchemy>=2.0.23
asyncpg
psycopg2-binary>=2.9.9
alembic>=1.12.1
python-dotenv>=1.0.0
pytest>=7.4.3
pytest-asyncio
pytest-cov>=4.1.0
httpx>=0.25.1
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
